+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 - Add expiremental `OnDiskDuplicateKeyTracker`
 - Add asn validator to j1Formats
+- Replace validateEntityWithSchema
 
 # 12.6.0 - 2024-04-24